fix: fixes for title and accept button label
fixes for title and accept button label
