+++ packages/ui-components/src/components/UITranslationInput/UITranslationButton.types.ts
@@ -1,4 +1,6 @@
-export interface TranslationButtonStrings {
+import type { UiIcons } from '../Icons';
+
+export interface TranslationInputStrings {
     acceptButtonLabel: string;
     cancelButtonLabel: string;
     i18nEntryExistsTooltip: string;
@@ -8,6 +10,7 @@ export interface TranslationButtonStrings {
     i18nValueMissingDescription: string;
     i18nReplaceWithExistingTooltip: string;
     i18nReplaceWithExistingDescription: string;
+    i18nEntryExistsInputTooltip: string;
 }
 
 export enum TranslationKeyGenerator {
@@ -37,25 +40,12 @@ export interface UITranslationProps {
     value?: string;
     id: string;
     disabled?: boolean;
-    // Existing I18n entries
-    entries: I18nBundle;
     // When entry exists in passed i18n entries and user clicked on show entry button
     onShowExistingEntry?: (entry: TranslationEntry) => void;
     // When creation of new i18n entry is requested
     onCreateNewEntry?: (entry: TranslationEntry) => void;
-    // PascalCase or camelCase to be used for key generation
-    // Default value is 'CamelCase'
-    namingConvention?: TranslationKeyGenerator;
     // Loader indicator
     busy?: UILoadButtonBusyProps;
-    // Default i18n prefix for "SingleBracketBinding"
-    i18nPrefix: string;
-    // Option to pass multiple allowed prefixes - if not passed then single "i18nPrefix" considered as allowed
-    allowedI18nPrefixes?: string[];
-    // Default pattern
-    defaultPattern: TranslationTextPattern;
-    // Allowed pattern
-    allowedPatterns: TranslationTextPattern[];
     // Opion to pass custom Texts for component's labels and tooltips
     strings?: TranslationButtonStrings;
 }
@@ -64,4 +54,23 @@ export interface UILoadButtonBusyProps {
     busy?: boolean;
     // If true set, then default time is 500ms
     useMinWaitingTime?: boolean | number;
+}
+
+export enum SuggestValueType {
+    Existing = 'Existing',
+    Update = 'Update',
+    New = 'New'
+}
+
+export interface TranslationSuggestValue {
+    entry: TranslationEntry;
+    icon?: UiIcons;
+    type: SuggestValueType;
+    i18n?: string;
+}
+
+export interface TranslationSuggest {
+    tooltip: string;
+    message?: React.ReactElement;
+    suggest?: TranslationSuggestValue;
 }
