+++ packages/ui-components/src/components/UITranslationInput/UITranslationInput.tsx
@@ -2,12 +2,132 @@ import React, { useCallback } from 'react';
 import type { ReactElement } from 'react';
 import { UITextInput } from '../UIInput';
 import type { ITextFieldProps } from '../UIInput';
+import { UiIcons } from '../Icons';
 import { UITranslationButton } from './UITranslationButton';
-import type { UITranslationProps } from './UITranslationButton.types';
+import type {
+    UITranslationProps,
+    TranslationSuggestValue,
+    TranslationSuggest,
+    I18nBundle,
+    TranslationTextPattern
+} from './UITranslationButton.types';
+import { TranslationKeyGenerator, SuggestValueType } from './UITranslationButton.types';
+import {
+    extractI18nKey,
+    generateI18nKey,
+    applyI18nPattern,
+    getTranslationByKey,
+    getTranslationByText
+} from './UITranslationUtils';
+import { defaultTranslationInputStrings } from './defaults';
+import { UIFormattedText, formatText } from './UIFormattedText';
 
 export interface UITranslationInputProps extends ITextFieldProps, UITranslationProps {
     id: string;
+    // Existing I18n entries
+    entries: I18nBundle;
+    // PascalCase or camelCase to be used for key generation
+    // Default value is 'CamelCase'
+    namingConvention?: TranslationKeyGenerator;
+    // Default i18n prefix for "SingleBracketBinding"
+    i18nPrefix: string;
+    // Option to pass multiple allowed prefixes - if not passed then single "i18nPrefix" considered as allowed
+    allowedI18nPrefixes?: string[];
+    // Default pattern
+    defaultPattern: TranslationTextPattern;
+    // Allowed pattern
+    allowedPatterns: TranslationTextPattern[];
 }
+
+/**
+ * Method returns suggestion object with message and tooltip based on passed translation button props.
+ *
+ * @param props Properties of translation input component.
+ * @returns Translation suggestion object.
+ */
+const getTranslationSuggestion = (props: UITranslationInputProps): TranslationSuggest => {
+    const {
+        value = '',
+        allowedPatterns,
+        entries,
+        strings = defaultTranslationInputStrings,
+        namingConvention = TranslationKeyGenerator.CamelCase,
+        defaultPattern,
+        i18nPrefix,
+        allowedI18nPrefixes
+    } = props;
+    const i18nKey = extractI18nKey(value, allowedPatterns, allowedI18nPrefixes || [i18nPrefix]);
+    let message = '';
+    let tooltip = '';
+    let suggest: TranslationSuggestValue;
+    if (i18nKey) {
+        // There is already i18n binding as value
+        const entry = getTranslationByKey(entries, i18nKey);
+        if (entry) {
+            tooltip = strings.i18nEntryExistsTooltip;
+            suggest = {
+                entry,
+                type: SuggestValueType.Existing,
+                icon: UiIcons.WorldArrow
+            };
+        } else {
+            message = strings.i18nKeyMissingDescription;
+            tooltip = strings.i18nKeyMissingTooltip;
+            suggest = {
+                entry: {
+                    key: {
+                        value: i18nKey
+                    },
+                    value: {
+                        value: i18nKey
+                    }
+                },
+                type: SuggestValueType.New,
+                icon: UiIcons.WorldWarning
+            };
+        }
+    } else {
+        // Use generation format passed from outside or use default as 'Standard';
+        const existingEntry = getTranslationByText(entries, value);
+        if (existingEntry) {
+            message = strings.i18nReplaceWithExistingDescription;
+            tooltip = strings.i18nReplaceWithExistingTooltip;
+            suggest = {
+                entry: existingEntry,
+                type: SuggestValueType.Update
+            };
+        } else {
+            message = strings.i18nValueMissingDescription;
+            tooltip = strings.i18nValueMissingTooltip;
+            const key = generateI18nKey(value, namingConvention, entries);
+            suggest = {
+                entry: {
+                    key: {
+                        value: key
+                    },
+                    value: {
+                        value
+                    }
+                },
+                type: SuggestValueType.New
+            };
+        }
+    }
+    // I18n string to apply for input value
+    suggest.i18n = applyI18nPattern(suggest.entry.key.value, defaultPattern, i18nPrefix);
+    // Format message to show in callout
+    const messageValues = {
+        key: suggest.entry.key.value,
+        value: suggest.entry.value.value,
+        i18n: suggest.i18n
+    };
+    tooltip = formatText(tooltip, messageValues);
+    return {
+        message: <UIFormattedText values={messageValues}>{message}</UIFormattedText>,
+        tooltip,
+        suggest
+    };
+};
 
 /**
  * Component to render translation input with button to provide helper callout with i18n generation option.
@@ -17,7 +137,6 @@ export interface UITranslationInputProps extends ITextFieldProps, UITranslationP
  */
 export function UITranslationInput(props: UITranslationInputProps): ReactElement {
     const {
-        title,
         id,
         className,
         onChange,
@@ -31,8 +150,11 @@ export function UITranslationInput(props: UITranslationInputProps): ReactElement
         namingConvention,
         onCreateNewEntry,
         onShowExistingEntry,
-        disabled
+        disabled,
+        strings
     } = props;
+
+    const suggestion = getTranslationSuggestion(props);
 
     let classNames = ' ui-translatable__input';
     // Custom external classes
@@ -46,23 +168,31 @@ export function UITranslationInput(props: UITranslationInputProps): ReactElement
         },
         [onChange]
     );
+    // Generate DOM id for i18n button
+    let buttonId = `${id}-i18n`;
+    let title = props.title;
+    if (suggestion.suggest?.type === SuggestValueType.Existing && strings?.i18nEntryExistsInputTooltip) {
+        // Change DOM id with additional suffix
+        buttonId += '-navigate';
+        // Overwrite title with information about reference entry
+        title = formatText(strings.i18nEntryExistsInputTooltip, {
+            value: value || '',
+            translation: suggestion.suggest.entry.value.value
+        });
+    }
 
     const onRenderSuffix = useCallback((): JSX.Element | null => {
         return (
             <UITranslationButton
-                id={`${id}-i18n`}
+                id={buttonId}
                 value={value}
-                allowedPatterns={allowedPatterns}
-                defaultPattern={defaultPattern}
-                entries={entries}
                 busy={busy}
-                i18nPrefix={i18nPrefix}
-                allowedI18nPrefixes={allowedI18nPrefixes}
-                namingConvention={namingConvention}
                 onCreateNewEntry={onCreateNewEntry}
                 onShowExistingEntry={onShowExistingEntry}
                 onUpdateValue={onUpdateValue}
                 disabled={disabled}
+                strings={strings}
+                suggestion={suggestion}
             />
         );
     }, [
@@ -89,3 +219,7 @@ export function UITranslationInput(props: UITranslationInputProps): ReactElement
         />
     );
 }
+
+UITranslationInput.defaultProps = {
+    strings: defaultTranslationInputStrings
+};
